fix Party on Death and Login status

+++ src/io/iologindata.cpp
@@ -66,27 +66,31 @@ void IOLoginData::setAccountType(uint32_t accountId, account::AccountType accoun
 	Database::getInstance().executeQuery(query.str());
 }
 
-void IOLoginData::updateOnlineStatus(uint32_t guid, bool login) {
-	// Verifique o status da conexão do player antes de atualizá-lo
-	bool isOnline = isPlayerOnline(guid);
-
-	if (login && !isOnline) {
-		std::ostringstream query;
-		query << "INSERT INTO `players_online` VALUES (" << guid << ')';
-		Database::getInstance().executeQuery(query.str());
-	} else if (!login && isOnline) {
-		std::ostringstream query;
-		query << "DELETE FROM `players_online` WHERE `player_id` = " << guid;
-		Database::getInstance().executeQuery(query.str());
-	}
-}
-
 bool IOLoginData::isPlayerOnline(uint32_t guid) {
 	std::ostringstream query;
 	query << "SELECT * FROM `players_online` WHERE `player_id` = " << guid;
 
 	DBResult_ptr result = Database::getInstance().storeQuery(query.str());
 	return (result && result->next());
+}
+
+void IOLoginData::updateOnlineStatus(uint32_t guid, bool login) {
+	bool isOnline = isPlayerOnline(guid);
+
+	if (login && isOnline) {
+		return;
+	}
+
+	std::ostringstream query;
+	if (login && !isOnline) {
+		query << "INSERT INTO `players_online` VALUES (" << guid << ')';
+	} else if (!login && isOnline) {
+		query << "DELETE FROM `players_online` WHERE `player_id` = " << guid;
+	} else {
+		query << "DELETE FROM `players_online` WHERE `player_id` = " << guid;
+	}
+
+	Database::getInstance().executeQuery(query.str());
 }
 
 bool IOLoginData::preloadPlayer(Player* player, const std::string &name) {
